- modified realsense435i.py to make sure the depth array is 16 bit - modified complete.py/manage.py to save depth and 'depth_array' type, so it can be handled specially in datastore.py - modified datastore.py to handle 'depth_array' as a 16bit grayscale PNG image.

+++ donkeycar/parts/realsense435i.py
@@ -131,8 +131,8 @@ class RealSense435i(object):
             depth_frame = aligned_frames.get_depth_frame() if aligned_frames is not None else frames.get_depth_frame()
             color_frame = aligned_frames.get_color_frame() if aligned_frames is not None else frames.get_color_frame()
 
-            # Convert images to numpy arrays
-            self.depth_image = np.asanyarray(depth_frame.get_data()) if self.enable_depth else None
+            # Convert depth to 16bit array, RGB into 8bit planar array
+            self.depth_image = np.asanyarray(depth_frame.get_data(), dtype=np.uint16) if self.enable_depth else None
             self.color_image = np.asanyarray(color_frame.get_data()) if self.enable_rgb else None
 
             if self.resize:
